rename resolver dir to graphql
