+++ projects/frontend/src/graphql/index.ts
