+++ projects/frontend/src/graphql/queries/artists-query.ts
