+++ projects/frontend/src/graphql/queries/genre-query.ts
