+++ projects/frontend/src/graphql/queries/song-types.ts
