+++ projects/frontend/src/graphql/types.local.ts
