+++ projects/frontend/src/graphql/types.ts
